backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.7.0
pydantic[email]==2.10.3
python-jose[cryptography]==3.3.0
httpx==0.28.1
websockets==14.1
